+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ UNRELEASED
 - Remove support for Django < 1.11.
 - Add ``AWS_SES_*`` settings to configure AWS credentials through
   ``settings.py``.
+- Add ``EmailBackend`` constructor arguments to override AWS credentials.
 
 1.0.0 - 2017-12-07
 ==================